# swagger_backend.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Inventory API",
    "description": "API for inventory and expense management",
    "version": "v1"
  },
  "paths": {
...
    "/api/Reports/profit-loss": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProfitLossReportDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfitLossReportDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfitLossReportDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reports/cash-flow": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowReportDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowReportDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowReportDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Reports/export": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "pdf"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CashFlowReportDto": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "cashFromSales": {
            "type": "number",
            "format": "double"
          },
          "otherIncome": {
            "type": "number",
            "format": "double"
          },
          "totalInflows": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "cashForExpenses": {
            "type": "number",
            "format": "double"
          },
          "cashForPurchases": {
            "type": "number",
            "format": "double"
          },
          "totalOutflows": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "netCashFlow": {
            "type": "number",
            "format": "double",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ProfitLossReportDto": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalRevenue": {
            "type": "number",
            "format": "double"
          },
          "costOfGoodsSold": {
            "type": "number",
            "format": "double"
          },
          "grossProfit": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "totalExpenses": {
            "type": "number",
            "format": "double"
          },
          "netProfit": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "expenseByCategory": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          },
          "revenueByProduct": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Enter JWT Token (e.g., 'Bearer 12345abcdef')",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}